Put in missing file editing step
This goes part of the way towards fixing #74.

After creating a new branch, no changes are made before trying to
commit (and the commit therefore fails, as can be seen in the
notebook). On the currently deployed pages, this works but only
because of an error in the previous step (see #79).

I have not rerun this because it will require doing everything from
the start, so it will need to be done at some point.