changed geocoder to Nominatim which uses OpenStreetMap
[Nominatum](https://nominatim.org/release-docs/develop/api/Overview/) doesn't have a limit but [AUP](https://operations.osmfoundation.org/policies/nominatim/) which states:

We are in principle happy for the public API to be used by external users for creative and unexpected uses. However, be aware that the service runs on donated servers and has a very limited capacity. We therefore ask you to limit your use and adhere to this usage policy.

Requirements:
No heavy uses (an absolute maximum of 1 request per second).
Provide a valid HTTP Referer or User-Agent identifying the application (stock User-Agents as set by http libraries will not do).
Clearly display attribution as suitable for your medium.
Data is provided under the ODbL license which requires to share alike (although small extractions are likely to be covered by fair usage / fair dealing).