Pin Jinja version to before breaking API change
We don't use Jinja directly but nbconvert does. Jinja version 3.1.x
has API changes (removing `contextfilter`) which don't work with
older nbconvert versions (nbconvert issue 1742). The newer version
of nbconvert should fix it, but we have pinned an older version
because of #186.
This is a temporary fix to get the build working again, but we
should eventually resolve the issuse in #186.